Rename useCurrentUser
